se agrego useRegionId
